fix sfx in dialog

+++ mm/src/code/z_message_nes.c
@@ -623,7 +623,7 @@ void Message_DrawTextNES(PlayState* play, Gfx** gfxP, u16 textDrawPos) {
             case 0x1E: // MESSAGE_SFX
                 if (((i + 1) == msgCtx->textDrawPos) && (msgCtx->msgMode == MSGMODE_TEXT_DISPLAYING)) {
                     stateTimerHi = msgCtx->decodedBuffer.schar[i + 1] << 8;
-                    stateTimerHi += msgCtx->decodedBuffer.schar[i + 2];
+                    stateTimerHi += (uint8_t)msgCtx->decodedBuffer.schar[i + 2];
                     Audio_PlaySfx(stateTimerHi);
                 }
                 if ((i + 1) == msgCtx->textDrawPos) {
@@ -1878,7 +1878,7 @@ void Message_DecodeNES(PlayState* play) {
                 (msgCtx->textBoxType == TEXTBOX_TYPE_8) || (msgCtx->textBoxType == TEXTBOX_TYPE_9) ||
                 (msgCtx->textBoxType == TEXTBOX_TYPE_B) || (msgCtx->unk11F0C == 3)) {
                 sfxHi = msgCtx->decodedBuffer.schar[decodedBufPos - 1] << 8;
-                sfxHi += msgCtx->decodedBuffer.schar[decodedBufPos];
+                sfxHi += (uint8_t)msgCtx->decodedBuffer.schar[decodedBufPos];
                 Audio_PlaySfx(sfxHi);
             }
         } else if (curChar == 0x1F) {
